% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions.R
\name{create_deck}
\alias{create_deck}
\title{Create a Deck of Playing Cards}
\usage{
create_deck(num = 1)
}
\arguments{
\item{num}{An integer specifying the number of decks to create (default is 1).}
}
\value{
A data frame representing a deck of playing cards with columns for values, suits, lookup, and points.
}
\description{
This function generates a deck of playing cards.
}
\details{
The deck consists of 52 cards, including numeric values from 2 to 10 and face cards (A, Q, K, J) for each of the four suits (Spades, Clubs, Diamonds, Hearts).
Each card has a unique identifier obtained by concatenating its value and suit.
Numeric values are assigned points equal to their face value, while face cards (Q, K, J) are assigned 10 points.
}
\examples{
create_deck()
create_deck(2)

}
