% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions.R
\name{did_dealer_win}
\alias{did_dealer_win}
\title{Check if the dealer has won.}
\usage{
did_dealer_win(dealer)
}
\arguments{
\item{dealer}{The cards held by the dealer.}
}
\value{
Logical value indicating whether the dealer has won.
}
\description{
This function checks if the dealer has won the game.
}
\examples{
deck <- create_deck()
dealer_hand <- dealer1(deck, player(deck))
did_dealer_win(dealer_hand)

}
