% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions.R
\name{hit_dealer}
\alias{hit_dealer}
\title{Hit the dealer with an additional card.}
\usage{
hit_dealer(dealer, deck)
}
\arguments{
\item{dealer}{The current cards held by the dealer.}

\item{deck}{The deck of cards.}
}
\value{
Updated dealer's hand after drawing a card.
}
\description{
This function gives an additional card to the dealer from the deck.
}
\examples{
deck <- create_deck()
dealer_hand <- dealer1(deck, player(deck))
dealer_hand <- hit_dealer(dealer_hand, deck)

}
