% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions.R
\name{hit_player}
\alias{hit_player}
\title{Hit the player with an additional card.}
\usage{
hit_player(player, deck)
}
\arguments{
\item{player}{The current cards held by the player.}

\item{deck}{The deck of cards.}
}
\value{
Updated player's hand after drawing a card.
}
\description{
This function gives an additional card to the player from the deck.
}
\examples{
deck <- create_deck()
player_hand <- player(deck)
player_hand <- hit_player(player_hand, deck)

}
