% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions.R
\name{keep_going_dealer}
\alias{keep_going_dealer}
\title{Check if the dealer should keep drawing cards.}
\usage{
keep_going_dealer(dealer)
}
\arguments{
\item{dealer}{The cards held by the dealer.}
}
\value{
Logical value indicating whether the dealer should keep drawing cards.
}
\description{
This function determines if the dealer should draw more cards.
}
\examples{
deck <- create_deck()
dealer_hand <- dealer1(deck, player(deck))
keep_going_dealer(dealer_hand)

}
