% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions.R
\name{keep_going_user}
\alias{keep_going_user}
\title{Check if the player should keep drawing cards.}
\usage{
keep_going_user(user)
}
\arguments{
\item{user}{The cards held by the player.}
}
\value{
Logical value indicating whether the player should keep drawing cards.
}
\description{
This function determines if the player should draw more cards.
}
\examples{
deck <- create_deck()
player_hand <- player(deck)
keep_going_user(player_hand)

}
