% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions.R
\name{player}
\alias{player}
\title{Deal cards to the player.}
\usage{
player(deck)
}
\arguments{
\item{deck}{The deck of cards.}
}
\value{
Cards dealt to the player.
}
\description{
This function deals cards to the player from the deck.
}
\examples{
deck <- create_deck()
player_hand <- player(deck)

}
