% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions.R
\name{update_deck_dealer}
\alias{update_deck_dealer}
\title{Update the deck after dealing cards to the dealer.}
\usage{
update_deck_dealer(deck, dealer)
}
\arguments{
\item{deck}{The original deck of cards.}

\item{dealer}{The cards dealt to the dealer.}
}
\value{
The updated deck after removing cards dealt to the dealer.
}
\description{
This function updates the deck by removing cards dealt to the dealer.
}
\examples{
deck <- create_deck()
dealer_hand <- dealer1(deck, player(deck))
updated_deck <- update_deck_dealer(deck, dealer_hand)

}
