% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions.R
\name{update_deck_user}
\alias{update_deck_user}
\title{Update the deck after dealing cards to the user.}
\usage{
update_deck_user(deck, player)
}
\arguments{
\item{deck}{The original deck of cards.}

\item{player}{The cards dealt to the player.}
}
\value{
The updated deck after removing cards dealt to the player.
}
\description{
This function updates the deck by removing cards dealt to the player.
}
\examples{
deck <- create_deck()
player_hand <- player(deck)
updated_deck <- update_deck_user(deck, player_hand)

}
