% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions.R
\name{update_points}
\alias{update_points}
\title{Update points based on the outcome of the game.}
\usage{
update_points(points, gamble, todo)
}
\arguments{
\item{points}{The current points of the player.}

\item{gamble}{The amount of points gambled.}

\item{todo}{The outcome of the game (win or loss).}
}
\value{
Updated points after considering the outcome of the game.
}
\description{
This function updates the player's points based on the game outcome.
}
\examples{
points <- 100
points <- update_points(points, 10, "win")

}
