% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions.R
\name{user_dealer_comparison}
\alias{user_dealer_comparison}
\title{Compare the user's and dealer's hands to determine the winner.}
\usage{
user_dealer_comparison(user, dealer)
}
\arguments{
\item{user}{The cards held by the user.}

\item{dealer}{The cards held by the dealer.}
}
\value{
Logical value indicating whether the user has won the game.
}
\description{
This function compares the user's and dealer's hands to determine the winner.
}
\examples{
deck <- create_deck()
user_hand <- player(deck)
dealer_hand <- dealer1(deck, player(deck))
user_dealer_comparison(user_hand, dealer_hand)

}
